add accept method - refacto unit test method for request with params
