+++ test/Events/index.test.js
@@ -1,5 +1,4 @@
-import {testGetter} from '../utils/http.js'
-import {testEventsGetter} from "../../src/Events/utils";
+import {testGetter, testParamsGetter} from '../utils/http.js'
 import Events from '../../src/Events'
 
 describe("Events#for", () =>  {
@@ -7,7 +6,7 @@ describe("Events#for", () =>  {
     return new Events(client).for("tata", {page: 1, from: 2})
   })
 
-  testEventsGetter("https://api.scalingo.com/v1/apps/tata/events", {page: 1, per_page: 2, from: 1},(client) =>  {
+  testParamsGetter("https://api.scalingo.com/v1/apps/tata/events", {page: 1, per_page: 2, from: 1},(client) =>  {
     return new Events(client).for("tata", {page: 1, per_page: 2, from: 1})
   })
 })
